wechat-imgs/requirements.txt
asn1crypto==0.21.1
attrs==16.3.0
Automat==0.5.0
beautifulsoup4==4.5.3
bs4==0.0.1
cffi==1.9.1
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
idna==2.5
incremental==16.10.1
itchat==1.2.30
lxml==3.7.3
numpy==1.12.0
olefile==0.44
packaging==16.8
parsel==1.1.0
Pillow==4.0.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
pywin32==220
queuelib==1.4.2
requests==2.13.0
Scrapy==1.3.3
service-identity==16.0.0
six==1.10.0
Twisted==17.1.0
w3lib==1.17.0
zope.interface==4.3.3
